Fixes delay param to proper type and converts to milliseconds

+++ jewelbots_nRF51822_board_v100/cores/JWB_nRF51822/LED.h
@@ -13,7 +13,7 @@ class LED {
 public:
   LED();
   ~LED();
-  void on(uint8_t number, char *color, uint8_t length);
+  void on(uint8_t number, char *color, uint32_t length);
   void color_lookup(char *color);
 
 };
